Addded support for keycloak and auth0
