+++ src/main/java/com/sum/security/service/AuthService.java
@@ -1,0 +1,5 @@
+package com.sum.security.service;
+
+public interface AuthService {
+    String fetchToken(String username, String password);
+}
