+++ src/main/resources/application-auth0.yml
@@ -1,19 +1,9 @@
-#server.ssl.key-store-password=tomcat
-#server.ssl.key-store=classpath:mykeystore.p12
-#server.ssl.key-store-type=PKCS12
-#server.ssl.key-alias=tomcat
-#server.port=8001
-#logging.level.org.springframework=DEBUG
-#server.use-forward-headers=true
-#server.servlet.context-path=/ldapauthentication
-#server.tomcat.use-relative-redirects=true
-
 server:
   tomcat:
     use-relative-redirects: true
 oauth:
-  clientId: C3UHISO3z60iF1JLG8L7VPUSWOASrJfO
-  clientSecret:
+  clientId: ${CLIENT_ID}
+  clientSecret: ${CLIENT_SECRET}
   realm: Username-Password-Authentication
   audience: https://skmaji.auth0.com/userinfo
   authorizeUrl: https://skmaji.auth0.com/authorize
