+++ src/main/resources/application-keycloak.yml
@@ -1,13 +1,3 @@
-#server.ssl.key-store-password=tomcat
-#server.ssl.key-store=classpath:mykeystore.p12
-#server.ssl.key-store-type=PKCS12
-#server.ssl.key-alias=tomcat
-#server.port=8001
-#logging.level.org.springframework=DEBUG
-#server.use-forward-headers=true
-#server.servlet.context-path=/ldapauthentication
-#server.tomcat.use-relative-redirects=true
-
 server:
   tomcat:
     use-relative-redirects: true
@@ -18,7 +8,7 @@ keycloak:
 oauth:
   clientId: ${CLIENT_ID}
   clientSecret: ${CLIENT_SECRET}
-  realm: GokDevelopers
+  realm: ${OAUTH_REALM}
   audience: ${keycloak.server.uri}/realms/${oauth.realm}/protocol/openid-connect/userinfo
   authorizeUrl: ${keycloak.server.uri}/realms/${oauth.realm}/protocol/openid-connect/auth
   tokenUrl: ${keycloak.server.uri}/realms/${oauth.realm}/protocol/openid-connect/token
